Upgrade neo project to net8

+++ neo/Cryptography/ECC/ECDsa.cs
@@ -29,7 +29,7 @@ namespace Neo.Cryptography.ECC
             BigInteger trunc = new BigInteger(message.Reverse().Concat(new byte[1]).ToArray());
             if (n.GetBitLength() < messageBitLength)
             {
-                trunc >>= messageBitLength - n.GetBitLength();
+                trunc >>= messageBitLength - (int) n.GetBitLength();
             }
             return trunc;
         }
@@ -49,7 +49,7 @@ namespace Neo.Cryptography.ECC
                     {
                         do
                         {
-                            k = rng.NextBigInteger(curve.N.GetBitLength());
+                            k = rng.NextBigInteger((int) curve.N.GetBitLength());
                         }
                         while (k.Sign == 0 || k.CompareTo(curve.N) >= 0);
                         ECPoint p = ECPoint.Multiply(curve.G, k);
@@ -72,7 +72,7 @@ namespace Neo.Cryptography.ECC
 
         private static ECPoint SumOfTwoMultiplies(ECPoint P, BigInteger k, ECPoint Q, BigInteger l)
         {
-            int m = Math.Max(k.GetBitLength(), l.GetBitLength());
+            int m = (int) Math.Max(k.GetBitLength(), l.GetBitLength());
             ECPoint Z = P + Q;
             ECPoint R = P.Curve.Infinity;
             for (int i = m - 1; i >= 0; --i)
@@ -143,7 +143,7 @@ namespace Neo.Cryptography.ECC
                 e = new BigInteger(msg.Reverse().Concat(new byte[1]).ToArray());
                 if (curve.N.GetBitLength() < messageBitLength)
                 {
-                    e >>= messageBitLength - curve.N.GetBitLength();
+                    e >>= messageBitLength - (int) curve.N.GetBitLength();
                 }
 
                 // step 1.6 Q = r^-1 (sR-eG)
